Pop-up Countries Menu Improvement
A 'stub' empty menu was appearing if the user cancelled
the password dialog for Tor. To fix this, only insert the
menu into the QAction once we've successfully authenticated
to Tor.

FIXME: The navigation bar layout doesn't get properly updated
when we call setManu() and layout()->update(). The menu arrow
is squash against the button icon a bit.